@ python-loader : replace deprecated np.float with np.float64 and update VERSION.txt file to correct version number.

+++ pyMCDS.py
@@ -203,7 +203,7 @@ class pyMCDS:
             a plane of voxel centers in the z-axis.
         Returns
         -------
-        conc_arr : array (np.float) shape=[nx_voxels, ny_voxels, nz_voxels]
+        conc_arr : array (np.float64) shape=[nx_voxels, ny_voxels, nz_voxels]
             Contains the concentration of the specified chemical in each voxel.
             The array spatially maps to a meshgrid of the voxel centers.
         """
@@ -354,15 +354,15 @@ class pyMCDS:
         # while we're at it, find the mesh
         coord_str = mesh_node.find('x_coordinates').text
         delimiter = mesh_node.find('x_coordinates').get('delimiter')
-        x_coords = np.array(coord_str.split(delimiter), dtype=np.float)
+        x_coords = np.array(coord_str.split(delimiter), dtype=np.float64)
 
         coord_str = mesh_node.find('y_coordinates').text
         delimiter = mesh_node.find('y_coordinates').get('delimiter')
-        y_coords = np.array(coord_str.split(delimiter), dtype=np.float)
+        y_coords = np.array(coord_str.split(delimiter), dtype=np.float64)
 
         coord_str = mesh_node.find('z_coordinates').text
         delimiter = mesh_node.find('z_coordinates').get('delimiter')
-        z_coords = np.array(coord_str.split(delimiter), dtype=np.float)
+        z_coords = np.array(coord_str.split(delimiter), dtype=np.float64)
 
         # reshape into a mesh grid
         xx, yy, zz = np.meshgrid(x_coords, y_coords, z_coords)
